Fix RMSF_FWHM header and move function to utils
