+++ RMtools_3D/do_RMsynth_3D.py
@@ -44,7 +44,7 @@ import astropy.io.fits as pf
 import astropy.table as at
 import numpy as np
 
-from RMutils.util_misc import interp_images
+from RMutils.util_misc import interp_images, remove_header_third_fourth_axis
 from RMutils.util_RM import do_rmsynth_planes, get_rmsf_planes
 
 if sys.version_info.major == 2:
@@ -474,12 +474,11 @@ def writefits(
             hduLst.writeto(fitsFileOut, output_verify="fix", overwrite=True)
             hduLst.close()
 
-    # Save the RMSF
+    # Save the RMSF: real, im, tot (4D) and FWHM (2D)
     if not not_rmsf:
-        ### TODO: do we want to keep current NAXIS=3 and remove 4th axis
-        ### or do we want to remove both 3rd and 4th axis AND SET NAXIS=2 ??
+        # Create header for RMSF_{real,imag,tot}.fits
 
-        # Put frequency axis first, and reshape to add degenerate axes:
+        # Put frequency axis first, and reshape to add degenerate Stokes axis:
         RMSFcube = np.reshape(RMSFcube, [RMSFcube.shape[0]] + output_axes)
         # Move Faraday depth axis to appropriate position to match header.
         RMSFcube = np.moveaxis(RMSFcube, 0, Ndim - freq_axis)
@@ -497,29 +496,12 @@ def writefits(
         )
         header["BUNIT"] = ""
 
-        rmheader = header.copy()
-        rmheader["BUNIT"] = "rad/m^2"
-        if "BTYPE" in rmheader:
-            del rmheader["BTYPE"]
-        # Because there can be problems with different axes having different FITS keywords,
-        # don't try to remove the FD axis, but just make it degenerate.
-        # Also requires np.expand_dims to set the correct NAXIS.
-        rmheader["NAXIS" + str(freq_axis)] = 1
-        rmheader["CTYPE" + str(freq_axis)] = (
-            "DEGENERATE",
-            "Axis left in to avoid FITS errors",
-        )
-        rmheader["CUNIT" + str(freq_axis)] = ""
-        rmheader["CRVAL" + str(freq_axis)] = 0  # doesnt mean anything
-        stokes_axis = None
-        for axis in range(1, rmheader["NAXIS"] + 1):
-            if "STOKES" in rmheader[f"CTYPE{axis}"]:
-                stokes_axis = axis
-        if stokes_axis is not None:
-            rmheader[f"CTYPE{stokes_axis}"] = (
-                "DEGENERATE",
-                "Axis left in to avoid FITS errors",
-            )
+        # Create header for RMSF_FHWM.fits
+        rmsffwhm_header = header.copy()
+        rmsffwhm_header["BUNIT"] = "rad/m^2"
+        rmsffwhm_header.pop("BTYPE", None)
+        # Remove 3rd and 4th axis, RMSF_FWHM is a plane, like the 2D peak maps
+        rmsffwhm_header = remove_header_third_fourth_axis(rmsffwhm_header)
 
         if write_seperate_FDF:  # more memory efficient as well
             header = _setStokes(header, "Q")
@@ -552,8 +534,9 @@ def writefits(
             del hdu2
             gc.collect()
 
+            print("shape fwhmRMSFCube",np.shape(fwhmRMSFCube))
             hdu3 = pf.PrimaryHDU(
-                np.expand_dims(fwhmRMSFCube.astype(dtFloat), axis=0), rmheader
+                np.expand_dims(fwhmRMSFCube.astype(dtFloat), axis=0), rmsffwhm_header
             )
             fitsFileOut = outDir + "/" + prefixOut + "RMSF_FWHM.fits"
             if verbose:
@@ -571,7 +554,7 @@ def writefits(
             del header["STOKES"]
             hdu2 = pf.ImageHDU(np.abs(RMSFcube).astype(dtFloat), header)
             hdu3 = pf.ImageHDU(
-                np.expand_dims(fwhmRMSFCube.astype(dtFloat), axis=0), rmheader
+                np.expand_dims(fwhmRMSFCube.astype(dtFloat), axis=0), rmsffwhm_header
             )
 
             fitsFileOut = outDir + "/" + prefixOut + "RMSF.fits"
